refactor: move snapshots/cmd_handler/ to snapshots/cli/cmd_handler/
Also move ShowResult from cmd_handler/show.py to snapshots/models.py
to avoid a circular import through snapshots/cli/__init__.py.
