+++ nbkp/snapshots/cli/show_cmd.py
@@ -11,7 +11,7 @@ import typer
 from ...clihelpers import OutputFormat
 from ...config.cli.helpers import load_config_or_exit, resolve_endpoints
 from ...config.epresolution import NetworkType
-from ..cmd_handler import show_all_syncs
+from .cmd_handler import show_all_syncs
 from ..output import print_human_show_results
 
 from ...disks.cli.helpers import managed_mount
